requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
easyocr==1.7.2
filelock==3.17.0
fsspec==2025.3.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
Levenshtein==0.27.1
MarkupSafe==3.0.2
mpmath==1.3.0
mss==10.0.0
networkx==3.2.1
ninja==1.11.1.3
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pip==25.0.1
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.6
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
setuptools==58.1.0
shapely==2.0.7
sympy==1.13.1
tifffile==2024.8.30
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
